+++ src/hooks/useScreenWidth.ts
@@ -1,0 +1,20 @@
+import { useState, useEffect } from 'react';
+
+function useScreenWidth() {
+  const [screenWidth, setScreenWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setScreenWidth(window.innerWidth);
+    };
+
+    window.addEventListener('resize', handleResize);
+    handleResize(); // 초기 로드 시 체크
+
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return screenWidth;
+}
+
+export default useScreenWidth;
